add a raw data transfer

+++ CANTRANSMIT/include/canTransmit.h
@@ -1,5 +1,5 @@
 #include "pico/stdlib.h"
 #include "MCP2515.h"
 namespace CAN{
-    int CanPreparation(can_frame &canFrame, uint8_t needToByte, float data);
+    int8_t CanPreparation(can_frame &canFrame, uint8_t needToByte, uint16_t data);
 }
